Added projects to example solution to ensure proper build
The project was previously configured to use the release builds of the dependent projects, which was a problem both for the automated build system but also for a user trying to get started with the project. Now, running the project will automatically build the referenced DLLs. Adding the projects into the example also allows the devleoper to understand what functionality is actually being called.