Updated code to handle waveforms with no burst timing
Some NR waveforms do not get created with any burst timing added to the TDMS file. The RFSGP Playback functions return null arrays in this case, which is an issue. This fix checks for null or empty arrays and sets the burst start/stop points based on the sample count.