Removed specific versions from references
Visual Studio's default behavior is to add references with "Specific Version" checked. This is rather frustrating for someone trying to open the code and use it in a variety of supported driver versions, so this setting has been set to false for all modules. Future build validation will include this.