Example Projects to Top of Solution (#133)
Moved example project to top for all example solutions that did not already place it there

As is documented in the discussion below, Visual Studio stores the startup project in the .suo files, which we do not include in our repository. Hence, when I ran the WLAN_dEVM example in C# on a fresh machine, it failed indicating that the startup project was set to the RFmxWLAN library. Hence, I had to manually change the startup project to the correct project.

Obviously, this is a big problem for our users. VS defaults to using the first project in the file as the startup. Hence, I verified or moved the example project to the top spot in each example solution file.

https://stackoverflow.com/questions/694730/why-is-set-as-startup-option-stored-in-the-suo-file-and-not-the-sln-file